♻️ Split test.py into a tests/ package
One module per public function (test_capacity/reserve/shrink_to_fit/allocated_bytes/stats), with the duplicated _pointer_size helper consolidated into tests/_support.py as POINTER_SIZE.

tests/__init__.py is required so unittest discovery recurses into the package; without it the suite silently collects zero tests on CPython 3.10/3.11.

Collateral: MANIFEST.in ships the package via recursive-include; cibuildwheel test-command pattern test.py -> 'test_*.py' (kept -s {project} so modules import as tests.* and the installed .so still outranks the namespace stub).
