+++ MANIFEST.in
@@ -1,5 +1,5 @@
 # Ship the test suite in the sdist so the extension can be built and tested
-# from a PyPI source download. The C source under src/ and the stub package are
-# already included by setuptools (ext_modules / package_data).
-include test.py
+# from a PyPI source download. The tests/ package is not part of the wheel
+# (packages = ["list_reserve"]), so the sdist is its only distribution channel.
+recursive-include tests *.py
 include LICENSE
