+++ tests/_support.py
@@ -1,0 +1,3 @@
+from sys import maxsize
+
+POINTER_SIZE = 8 if maxsize > 2**32 else 4
